refactor : NewStoreRequestViewModel Dependency 리팩토링 (#286)
